pyproject.toml
[tool.poetry]
name = "azuma"
version = "0.1.0"
description = "Yet another Sigma library for Python"
authors = ["Manabu Niseki <manabu.niseki@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
regex = ">=2022"
lark-parser = ">=0.12,<1.0"
pyyaml = ">=6.0,<7.0"
pydantic = ">=1.8,<2.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
gitpython = "^3.1.30"
loguru = "^0.6.0"
mypy = "^0.991"
pre-commit = "^3.0.1"
pytest = "^7.2.1"
pytest-pretty = "^1.0.1"
pytest-randomly = "^3.12.0"
pyupgrade = "^3.3.1"
ruff = "^0.0.236"
types-pyyaml = "^6.0.12.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "T20",  # flake8-print
]
ignore = [
    "E501",  # line too long, handled by black
]
